+++ jgclark.DashboardReact/src/react/components/ItemGrid.jsx
@@ -34,18 +34,18 @@ function ItemGrid({ items, thisSection }: Props): React$Node {
 
  // Effect to handle item processing
 useEffect(() => {
-  if (!reactSettings) return;
-  if (!(thisSection.sectionCode === 'OVERDUE')) return;
-  if (!reactSettings.overdueProcessing) return;
-  logDebug('ItemGrid', 'effect 2', reactSettings);
-  const currentOverdueIndex = reactSettings.currentOverdueIndex || 0;
-  const dialogIsOpen = reactSettings.dialogData?.isOpen || false;
+  if (!reactSettings) return
+  if (!(thisSection.sectionCode === 'OVERDUE')) return
+  if (!reactSettings.overdueProcessing) return
+  logDebug('ItemGrid', 'effect 2', reactSettings)
+  const currentOverdueIndex = reactSettings.currentOverdueIndex || 0
+  const dialogIsOpen = reactSettings.dialogData?.isOpen || false
 
   // Check to advance to the next item
   // We know the next item is ready to be processed because the dialog was closed
-  logDebug('ItemGrid', `currentOverdueIndex: ${currentOverdueIndex}, dialogIsOpen: ${dialogIsOpen}`, itemsCopy[currentOverdueIndex]);
+  logDebug('ItemGrid', `currentOverdueIndex: ${currentOverdueIndex}, dialogIsOpen: ${dialogIsOpen}`, itemsCopy[currentOverdueIndex])
   if (!dialogIsOpen && currentOverdueIndex < itemsCopy.length && itemsCopy[currentOverdueIndex]) {
-    logDebug('ItemGrid', `Opening next installment ${currentOverdueIndex}`);
+    logDebug('ItemGrid', `Opening next installment ${currentOverdueIndex}`)
     // Open dialog with the next item's details
     setReactSettings((prev) => ({
       ...prev,
@@ -54,9 +54,9 @@ useEffect(() => {
         details: { item: itemsCopy[currentOverdueIndex] },
       },
       currentOverdueIndex: currentOverdueIndex + 1, // Increment the index here
-    }));
+    }))
   } else if (!dialogIsOpen && currentOverdueIndex >= itemsCopy.length) {
-    logDebug('ItemGrid', `Over the limit?  ${currentOverdueIndex >= itemsCopy.length} ${currentOverdueIndex} >= ${itemsCopy.length}`);
+    logDebug('ItemGrid', `Over the limit?  ${currentOverdueIndex >= itemsCopy.length} ${currentOverdueIndex} >= ${itemsCopy.length}`)
     if (itemsCopy?.length > 0) {
       // on first load, we don't want to reset yet
       // All items processed, reset the processing state
@@ -65,14 +65,14 @@ useEffect(() => {
         overdueProcessing: false,
         currentOverdueIndex: -1,
         dialogData: { isOpen: false, details: null },
-      }));
-      setItemsCopy([]); // Clear itemsCopy to allow for reinitialization later
+      }))
+      setItemsCopy([]) // Clear itemsCopy to allow for reinitialization later
     }
   }
-}, [thisSection.sectionCode, reactSettings, itemsCopy]);
+}, [thisSection.sectionCode, reactSettings, itemsCopy])
 
 
-  const visibleItems = items.map((item, index) => <ItemRow key={index} item={item} thisSection={thisSection} />)
+  const visibleItems = items.map((item, index) => <ItemRow key={item.ID} item={item} thisSection={thisSection} />)
 
   return (
     <div className="sectionItemsGrid" id={`${thisSection.ID}-Section`}>